Updated xrengine-matchmaking Helm chart to work.
director and matchmaking code was updated to create a liveness/readiness
webserver so that the Helm deployment can properly monitor them. They
also source some variables, like the matchmaking function port, from
environment variables that the Helm deployment sets.

Made a Helm chart for the Open Match deployment that includes an ingress
for the frontend service, so that kubectl port-forward isn't necessary, and
so that a production deployment has an ingress for that traffic.

xrengine-matchmaking Helm chart updated to work. Removed frontend templates
since they're not needed. Put all of the matchmaking resources in their own
namespace.